requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.1.5
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
executing==2.0.1
idna==3.9
inflection==0.5.1
ipython==8.27.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
